Fix module type commonjs
Now our build exports a commonjs module we need to tell the package.json this is the case. This fixes that. Unfortunately that change breaks storybook. I've then switched storybook to use their new bable configuration mechanism https://storybook.js.org/docs/react/configure/babel and configured the modules toalways be commonjs. This feels like a bit of a hack but I think it works.